+++ lib/Misc/NoteUtils.php
@@ -7,44 +7,79 @@ class NoteUtils{
     }
     public function getMetadata($carnetFolder, $path){
         $meta = array();
-        $tmppath = tempnam(sys_get_temp_dir(), uniqid().".zip");
         $node = $carnetFolder->get($path);
-        
-        file_put_contents($tmppath, $node->fopen("r"));
-            $zipFile = new \PhpZip\ZipFile();
-            $zipFile->openFromStream(fopen($tmppath, "r")); //issue with encryption when open directly + unexpectedly faster to copy before Oo'
+        if($node->getType() === "dir"){
             $meta['lastmodfile'] = $node->getMTime();
             try{
-                $meta['metadata'] = json_decode($zipFile->getEntryContents("metadata.json"));
-            } catch(\PhpZip\Exception\ZipNotFoundEntry $e){
+                $meta['metadata'] = json_decode($node->get('metadata.json')->getContent());
+            } catch(\OCP\Files\NotFoundException $e){
             
             }
             try{
-            
-            $meta['shorttext'] = self::getShortTextFromHTML($zipFile->getEntryContents("index.html"));
-            $meta['media'] = array();
-            $meta['previews'] = array();
+                
+                $meta['shorttext'] = self::getShortTextFromHTML($node->get('index.html')->getContent());
+                $meta['media'] = array();
+                $meta['previews'] = array();
 
-            $i=0;
-            try{
-                foreach($zipFile->getListFiles() as $f){
-                    if(substr($f, 0, strlen("data/preview")) === "data/preview"){
-                        $meta['previews'][$i] = "./note/getmedia?note=".$path."&media=".$f;
-                        $i++;
-                    } else if(substr($f, 0, strlen("data/")) === "data/") {
-                        $meta['media'][$i] = "./note/getmedia?note=".$path."&media=".$f;
+                $i=0;
+                try{
+                    foreach($node->get('data')->getDirectoryListing() as $in){
+                        if(substr($in->getName(), 0, strlen("preview")) === "preview"){
+                            $meta['previews'][$i] = "./note/getmedia?note=".$path."&media=data/".$in->getName();
+                            $i++;
+                        } else  {
+                            $meta['media'][$i] = "./note/getmedia?note=".$path."&media=data/".$in->getName();
+                        }
+                        
                     }
+                }
+                catch(\OCP\Files\NotFoundException$e){
                     
                 }
             }
-            catch(\PhpZip\Exception\ZipNotFoundEntry $e){
+            catch(\OCP\Files\NotFoundException $e){
+                    
+            }
+        }
+        else{
+            $tmppath = tempnam(sys_get_temp_dir(), uniqid().".zip");
+            
+            file_put_contents($tmppath, $node->fopen("r"));
+                $zipFile = new \PhpZip\ZipFile();
+                $zipFile->openFromStream(fopen($tmppath, "r")); //issue with encryption when open directly + unexpectedly faster to copy before Oo'
+                $meta['lastmodfile'] = $node->getMTime();
+                try{
+                    $meta['metadata'] = json_decode($zipFile->getEntryContents("metadata.json"));
+                } catch(\PhpZip\Exception\ZipNotFoundEntry $e){
+                
+                }
+                try{
+                
+                $meta['shorttext'] = self::getShortTextFromHTML($zipFile->getEntryContents("index.html"));
+                $meta['media'] = array();
+                $meta['previews'] = array();
+
+                $i=0;
+                try{
+                    foreach($zipFile->getListFiles() as $f){
+                        if(substr($f, 0, strlen("data/preview")) === "data/preview"){
+                            $meta['previews'][$i] = "./note/getmedia?note=".$path."&media=".$f;
+                            $i++;
+                        } else if(substr($f, 0, strlen("data/")) === "data/") {
+                            $meta['media'][$i] = "./note/getmedia?note=".$path."&media=".$f;
+                        }
+                        
+                    }
+                }
+                catch(\PhpZip\Exception\ZipNotFoundEntry $e){
+                    
+                }
+            } catch(\PhpZip\Exception\ZipNotFoundEntry $e){
+                $meta['shorttext'] = "";
                 
             }
-        } catch(\PhpZip\Exception\ZipNotFoundEntry $e){
-            $meta['shorttext'] = "";
-            
+            unlink($tmppath);
         }
-        unlink($tmppath);
         return $meta;
         
     }
